backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
qrcode[pil]>=7.4.2
pillow>=10.1.0
faker>=20.1.0
pandas>=2.1.4
numpy>=1.25.2
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0
fastapi-cors>=0.0.6
python-dateutil>=2.8.2
google-generativeai>=0.3.2
